update in datatype as BigDecimal

+++ src/main/java/com/product/ecomproducts/model/Product.java
@@ -22,7 +22,7 @@ public class Product {
     private String id;
     private String name;
     private String description;
-    private Integer price;
+    private BigDecimal price;
 
 
 }
